+++ plugins/gui/include/gui/graph_tab_widget/graph_tab_widget.h
@@ -45,6 +45,7 @@ namespace hal
     class GraphTabWidget : public ContentWidget
     {
         Q_OBJECT
+        Q_PROPERTY(bool darkCursor READ darkCursor WRITE setDarkCursor);
 
     public:
         /**
@@ -83,6 +84,16 @@ namespace hal
          * GraphicsItem can be seen.
          */
         void ensureSelectionVisible();
+
+        /**
+         * Getter for darkCursor property used for module selection
+         */
+        bool darkCursor() const;
+
+        /**
+         * Setter for darkCursor property used for module selection
+         */
+        void setDarkCursor(bool dark);
 
         enum KeyboardModifier{Alt, Ctrl, Shift};
         Q_ENUM(KeyboardModifier)
@@ -142,6 +153,13 @@ namespace hal
          */
         void handleModuleFocus(u32 moduleId);
 
+        /**
+         * Change shape of cursor to indicate that a module should be picked by user
+         *
+         * @param on - true=on,  false=off
+         */
+        void setModuleSelectCursor(bool on);
+
     private:
         QTabWidget* mTabWidget;
         QVBoxLayout* mLayout;
@@ -172,5 +190,7 @@ namespace hal
         static bool initSettings();
 
         QMap<KeyboardModifier, Qt::KeyboardModifier> mKeyModifierMap;
+        bool mModuleSelectCursor;
+        bool mDarkCursor;
     };
 }