+++ plugins/gui/src/graph_tab_widget/graph_tab_widget.cpp
@@ -12,6 +12,8 @@
 #include <QVBoxLayout>
 #include <QShortcut>
 #include <QKeySequence>
+#include <QCursor>
+#include <QBitmap>
 
 namespace hal
 {
@@ -72,7 +74,9 @@ namespace hal
         return true;
     }
 
-    GraphTabWidget::GraphTabWidget(QWidget* parent) : ContentWidget("Graph-Views", parent), mTabWidget(new QTabWidget()), mLayout(new QVBoxLayout()), mZoomFactor(1.2)
+    GraphTabWidget::GraphTabWidget(QWidget* parent) : ContentWidget("Graph-Views", parent),
+        mTabWidget(new QTabWidget()), mLayout(new QVBoxLayout()), mZoomFactor(1.2),
+        mModuleSelectCursor(false), mDarkCursor(false)
     {
         mContentLayout->addWidget(mTabWidget);
         mTabWidget->setTabsClosable(true);
@@ -249,5 +253,38 @@ namespace hal
             }
         }
         return -1;
+    }
+    bool GraphTabWidget::darkCursor() const
+    {
+        return mDarkCursor;
+    }
+
+    void GraphTabWidget::setDarkCursor(bool dark)
+    {
+        if (dark==mDarkCursor) return;
+        mDarkCursor = dark;
+        setModuleSelectCursor(mModuleSelectCursor);
+    }
+
+    void GraphTabWidget::setModuleSelectCursor(bool on)
+    {
+        mModuleSelectCursor = on;
+        int n = mTabWidget->count();
+        if (mModuleSelectCursor)
+        {
+            QBitmap bitmap(":/icons/module_cursor","PNG");
+            QCursor modCurs(bitmap,bitmap);
+            if (!mDarkCursor)
+            {
+                QBitmap filled(bitmap);
+                filled.fill(Qt::white);
+                modCurs = QCursor(filled,bitmap);
+            }
+            for (int i=0; i<n; i++)
+                mTabWidget->widget(i)->setCursor(modCurs);
+        }
+        else
+            for (int i=0; i<n; i++)
+                mTabWidget->widget(i)->unsetCursor();
     }
 }